// Atspm/WebUI/src/features/charts/leftTurnGapAnalysis/types.ts
// #region license
// Copyright 2026 Utah Departement of Transportation
// for WebUI - types.ts
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//http://www.apache.org/licenses/LICENSE-2.
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
// #endregion
import { LeftTurnGapAnalysisResult } from '@/api/reports'
import { BaseChartOptions, ChartType } from '@/features/charts/common/types'

export interface LeftTurnGapAnalysisChartOptions extends BaseChartOptions {
  binSize: number
  gap1Min: number
  gap1Max: number
  gap2Min: number
  gap2Max: number
  gap3Min: number
  gap3Max: number
  gap4Min: number
  trendLineGapThreshold: number
}

export interface LeftTurnGapAnalysisChartOptionsDefaults {
  binSize: { id: number; value: string; option: string }
  gap1Min: { id: number; value: string; option: string }
  gap1Max: { id: number; value: string; option: string }
  gap2Min: { id: number; value: string; option: string }
  gap2Max: { id: number; value: string; option: string }
  gap3Min: { id: number; value: string; option: string }
  gap3Max: { id: number; value: string; option: string }
  gap4Min: { id: number; value: string; option: string }
  trendLineGapThreshold: { id: number; value: string; option: string }
}

export type RawLeftTurnGapData = LeftTurnGapAnalysisResult

export interface RawLeftTurnGapAnalysisResponse {
  type: ChartType.LeftTurnGapAnalysis
  data: RawLeftTurnGapData[]
}